+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorTestUtil.kt
@@ -32,12 +32,16 @@ import java.io.StringWriter
 
 object CodeGeneratorTestUtil {
 
+    @Suppress("ConstPropertyName")
     const val packageName = "net.pwall.json.schema.codegen.test.kotlin"
     val packageDirs = packageName.split('.')
     val dirs = listOf("com", "example")
 
     fun resultFile(filename: String): String =
         File("src/test/kotlin/net/pwall/json/schema/codegen/test/kotlin/$filename.kt").reader().readText()
+
+    fun resultJava(filename: String): String =
+        File("src/test/kotlin/net/pwall/json/schema/codegen/test/java/$filename.java").reader().readText()
 
     fun createHeader(name: String, comment: String? = null) =
 """/*